+++ fastats/linear_algebra/__init__.py
@@ -1,0 +1,9 @@
+
+from fastats.linear_algebra.lu import lu, lu_inplace
+from fastats.linear_algebra.pca import pca
+
+__all__ = [
+    'lu',
+    'lu_inplace',
+    'pca',
+]