+++ fastats/scaling/__init__.py
@@ -1,0 +1,11 @@
+
+from fastats.scaling.scaling import standard, min_max, rank, scale, demean
+
+
+__all__ = [
+    'standard',
+    'min_max',
+    'rank',
+    'scale',
+    'demean'
+]