+++ fastats/utilities/__init__.py
@@ -1,0 +1,6 @@
+
+from fastats.utilities.arrays import is_square
+
+__all__ = [
+    'is_square'
+]