+++ tests/maths/test_sum_sq_dev.py
@@ -35,4 +35,4 @@ def test_nan_values():
 
 if __name__ == '__main__':
     import pytest
-    pytest.main()
+    pytest.main([__file__])