+++ releases.md
@@ -3,7 +3,14 @@
 
 #### New features
 
+- newton_raphson : Root finding using the Newton-Raphson iteration method
 - erf : error function using Abramowitz and Stegun method (maximum error: 1.5e-7) 
+- correlation : Spearman and Pearson correlation coefficient functions
+- pca : Principal Component Analysis, returning the transformed data, not eigenvalues
+or eigenvectors.
+- Added Windows CI builds: #23
+- LU Decomposition
+- binary_search : Root finding using the bisection method
 
 #### Bug fixes